backend/pyproject.toml
[tool.poetry]
name = "entzun-backend"
version = "0.1.0"
description = "FastAPI backend for audio transcription and report generation"
authors = ["Endika Iglesias <endika2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
langchain = "^0.1.12"
langchain-openai = "^0.1.5"
openai-whisper = "^20231117"
python-multipart = "^0.0.9"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
openai = "^1.30.1"
tiktoken = "^0.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80
target-version = "py312"
select = ["E", "F", "I", "B", "W", "C90"]
ignore = []

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "langchain", "openai"]
